[fix] Add 'ca' setting when 'sslCAPath' is present
Mongo driver has an issue connecting to MongoS server (see
https://www.compose.com/articles/one-missing-key-and-how-it-broke-node-js-and-mongodb/),
which makes it necessary to have a 'ca' option if we have the 'sslCA'
when connecting to server.

This might be fixed on future versions of the driver, but for now
we need to workaround the issue until we can use a fixed version of the
driver.